+++ pom.xml
@@ -10,7 +10,7 @@
 	</parent>
 	<groupId>io.github.mihaita-tinta</groupId>
 	<artifactId>webauthn-spring-boot-starter</artifactId>
-	<version>0.1.1-RELEASE</version>
+	<version>0.2.0-SNAPSHOT</version>
 	<packaging>jar</packaging>
 	<name>webauthn-spring-boot-starter</name>
 	<description>Simple Webauthn starter project</description>
@@ -112,9 +112,10 @@
 			<scope>test</scope>
 		</dependency>
 		<dependency>
-			<groupId>org.springframework.boot</groupId>
-			<artifactId>spring-boot-starter-data-jpa</artifactId>
-			<scope>provided</scope>
+			<groupId>io.github.mihaita-tinta</groupId>
+			<artifactId>webauthn-spring-boot-starter-jpa</artifactId>
+			<version>${project.version}</version>
+			<scope>test</scope>
 		</dependency>
 		<dependency>
 			<groupId>com.h2database</groupId>
