+++ src/main/java/io/github/webauthn/domain/WebAuthnCredentialsInMemoryRepository.java
@@ -78,4 +78,12 @@ public class WebAuthnCredentialsInMemoryRepository implements WebAuthnCredential
                 .findFirst();
     }
 
+    @Override
+    public WebAuthnCredentials save(byte[] credentialId, Long appUserId, Long count, byte[] publicKeyCose, String userAgent) {
+
+        return new DefaultWebAuthnCredentials(credentialId,
+                appUserId, count, publicKeyCose,userAgent
+        );
+    }
+
 }
