+++ src/main/java/io/github/webauthn/domain/WebAuthnCredentialsRepository.java
@@ -1,20 +1,24 @@
 package io.github.webauthn.domain;
 
 
+import io.github.webauthn.dto.RegistrationStartRequest;
+
 import java.util.List;
 import java.util.Optional;
 
-public interface WebAuthnCredentialsRepository {
+public interface WebAuthnCredentialsRepository<T extends WebAuthnCredentials> {
 
-    List<WebAuthnCredentials> findAllByAppUserId(Long userId);
+    List<T> findAllByAppUserId(Long userId);
 
-    Optional<WebAuthnCredentials> findByCredentialIdAndAppUserId(byte[] credentialId, Long userId);
+    Optional<T> findByCredentialIdAndAppUserId(byte[] credentialId, Long userId);
 
-    List<WebAuthnCredentials> findByCredentialId(byte[] credentialId);
+    List<T> findByCredentialId(byte[] credentialId);
 
-    WebAuthnCredentials save(WebAuthnCredentials credentials);
+    T save(T credentials);
 
     void deleteByAppUserId(Long appUserId);
 
     void deleteById(Long id);
+
+    T save(byte[] credentialId, Long appUserId, Long count, byte[] publicKeyCose, String userAgent);
 }
