+++ src/main/resources/META-INF/spring.factories
@@ -1,6 +1,5 @@
 # Auto Configure
 org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-io.github.webauthn.WebAuthnSpringDataAutoConfiguration,\
 io.github.webauthn.WebAuthnInMemoryAutoConfiguration,\
 io.github.webauthn.WebAuthnConfig,\
 io.github.webauthn.webflux.WebAuthnWebFluxConfig
