+++ src/test/java/io/github/webauthn/flows/WebAuthnRegistrationStartStrategyTest.java
@@ -5,11 +5,10 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.yubico.webauthn.data.ByteArray;
 import io.github.webauthn.BytesUtil;
 import io.github.webauthn.JsonConfig;
-import io.github.webauthn.domain.WebAuthnCredentials;
-import io.github.webauthn.domain.WebAuthnCredentialsRepository;
-import io.github.webauthn.domain.WebAuthnUser;
-import io.github.webauthn.domain.WebAuthnUserRepository;
+import io.github.webauthn.domain.*;
 import io.github.webauthn.dto.RegistrationStartRequest;
+import io.github.webauthn.jpa.JpaWebAuthnCredentials;
+import io.github.webauthn.jpa.JpaWebAuthnUser;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
@@ -19,8 +18,8 @@ import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.security.test.context.support.WithMockUser;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.transaction.annotation.Transactional;
 
+import javax.transaction.Transactional;
 import java.time.LocalDateTime;
 import java.util.Base64;
 
@@ -58,10 +57,10 @@ public class WebAuthnRegistrationStartStrategyTest {
         RegistrationStartRequest request = new RegistrationStartRequest();
         request.setUsername("newjunit");
         this.mockMvc.perform(
-                post("/registration/start")
-                        .accept(MediaType.APPLICATION_JSON)
-                        .content(mapper.writeValueAsString(request))
-                        .contentType(MediaType.APPLICATION_JSON))
+                        post("/registration/start")
+                                .accept(MediaType.APPLICATION_JSON)
+                                .content(mapper.writeValueAsString(request))
+                                .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isOk())
                 .andDo(document("registration-start-new-user"));
     }
@@ -72,10 +71,10 @@ public class WebAuthnRegistrationStartStrategyTest {
         RegistrationStartRequest request = new RegistrationStartRequest();
         request.setRegistrationAddToken("token-123");
         this.mockMvc.perform(
-                post("/registration/start")
-                        .accept(MediaType.APPLICATION_JSON)
-                        .content(mapper.writeValueAsString(request))
-                        .contentType(MediaType.APPLICATION_JSON))
+                        post("/registration/start")
+                                .accept(MediaType.APPLICATION_JSON)
+                                .content(mapper.writeValueAsString(request))
+                                .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isBadRequest())
                 .andDo(document("registration-start-add-device-invalid-token"));
     }
@@ -86,10 +85,10 @@ public class WebAuthnRegistrationStartStrategyTest {
         RegistrationStartRequest request = new RegistrationStartRequest();
         request.setRegistrationAddToken(Base64.getEncoder().encodeToString("token-123".getBytes()));
         this.mockMvc.perform(
-                post("/registration/start")
-                        .accept(MediaType.APPLICATION_JSON)
-                        .content(mapper.writeValueAsString(request))
-                        .contentType(MediaType.APPLICATION_JSON))
+                        post("/registration/start")
+                                .accept(MediaType.APPLICATION_JSON)
+                                .content(mapper.writeValueAsString(request))
+                                .contentType(MediaType.APPLICATION_JSON))
                 .andExpect(status().isBadRequest())
                 .andDo(document("registration-start-add-device-not-found-token"));
     }
@@ -99,7 +98,7 @@ public class WebAuthnRegistrationStartStrategyTest {
         byte[] bytes = "token-123".getBytes();
         String registrationAddToken = Base64.getEncoder().encodeToString(bytes);
 
-        WebAuthnUser user = new WebAuthnUser();
+        JpaWebAuthnUser user = new JpaWebAuthnUser();
         user.setUsername("junit");
         user.setAddToken(bytes);
         user.setRegistrationAddStart(LocalDateTime.now().minusMinutes(1));
@@ -108,10 +107,10 @@ public class WebAuthnRegistrationStartStrategyTest {
         RegistrationStartRequest request = new RegistrationStartRequest();
         request.setRegistrationAddToken(registrationAddToken);
         this.mockMvc.perform(
-                post("/registration/start")
-                        .accept(MediaType.APPLICATION_JSON)
-                        .content(mapper.writeValueAsString(request))
-                        .contentType(MediaType.APPLICATION_JSON))
+                        post("/registration/start")
+                                .accept(MediaType.APPLICATION_JSON)
+                                .content(mapper.writeValueAsString(request))
+                                .contentType(MediaType.APPLICATION_JSON))
                 .andDo(print())
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.status").value("OK"))
@@ -127,18 +126,18 @@ public class WebAuthnRegistrationStartStrategyTest {
         byte[] bytes = "token-123".getBytes();
         String token = Base64.getEncoder().encodeToString(bytes);
 
-        WebAuthnUser user = new WebAuthnUser();
+        JpaWebAuthnUser user = new JpaWebAuthnUser();
         user.setUsername("junit");
         user.setRecoveryToken(bytes);
-        user = webAuthnUserRepository.save(user);
+        webAuthnUserRepository.save(user);
 
         RegistrationStartRequest request = new RegistrationStartRequest();
         request.setRecoveryToken(token);
         this.mockMvc.perform(
-                post("/registration/start")
-                        .accept(MediaType.APPLICATION_JSON)
-                        .content(mapper.writeValueAsString(request))
-                        .contentType(MediaType.APPLICATION_JSON))
+                        post("/registration/start")
+                                .accept(MediaType.APPLICATION_JSON)
+                                .content(mapper.writeValueAsString(request))
+                                .contentType(MediaType.APPLICATION_JSON))
                 .andDo(print())
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.status").value("OK"))
@@ -155,10 +154,10 @@ public class WebAuthnRegistrationStartStrategyTest {
         RegistrationStartRequest request = new RegistrationStartRequest();
         request.setRecoveryToken(Base64.getEncoder().encodeToString("token-321".getBytes()));
         this.mockMvc.perform(
-                post("/registration/start")
-                        .accept(MediaType.APPLICATION_JSON)
-                        .content(mapper.writeValueAsString(request))
-                        .contentType(MediaType.APPLICATION_JSON))
+                        post("/registration/start")
+                                .accept(MediaType.APPLICATION_JSON)
+                                .content(mapper.writeValueAsString(request))
+                                .contentType(MediaType.APPLICATION_JSON))
                 .andDo(print())
                 .andExpect(status().isBadRequest())
                 .andDo(document("registration-start-recovery-invalid"));
@@ -168,21 +167,21 @@ public class WebAuthnRegistrationStartStrategyTest {
     @WithMockUser("junit")
     public void testRegisterCredentialsForExistingUser() throws Exception {
 
-        WebAuthnUser user = new WebAuthnUser();
+        JpaWebAuthnUser user = new JpaWebAuthnUser();
         user.setUsername("junit");
         webAuthnUserRepository.save(user);
 
-        WebAuthnCredentials credentials = new WebAuthnCredentials();
+        JpaWebAuthnCredentials credentials = new JpaWebAuthnCredentials();
         credentials.setAppUserId(user.getId());
         credentials.setCredentialId(BytesUtil.longToBytes(123L));
         credentialsRepository.save(credentials);
 
         RegistrationStartRequest request = new RegistrationStartRequest();
         this.mockMvc.perform(
-                post("/registration/start")
-                        .accept(MediaType.APPLICATION_JSON)
-                        .content(mapper.writeValueAsString(request))
-                        .contentType(MediaType.APPLICATION_JSON))
+                        post("/registration/start")
+                                .accept(MediaType.APPLICATION_JSON)
+                                .content(mapper.writeValueAsString(request))
+                                .contentType(MediaType.APPLICATION_JSON))
                 .andDo(print())
                 .andExpect(status().is2xxSuccessful())
                 .andDo(document("registration-start-existing-user"));
